Recalculate relation indices on commit in schema transactions. Ignore indices in schema tx reads. First fix iteration
