+++ compiler/transformation/transform.rs
@@ -3,8 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-
-use concept::type_::type_manager::TypeManager;
+use concept::{thing::thing_manager::ThingManager, type_::type_manager::TypeManager};
 use storage::snapshot::ReadableSnapshot;
 
 use crate::{
@@ -20,14 +19,14 @@ use crate::{
 
 pub fn apply_transformations(
     snapshot: &impl ReadableSnapshot,
-    type_manager: &TypeManager,
+    thing_manager: &ThingManager,
     pipeline: &mut AnnotatedPipeline,
 ) -> Result<(), StaticOptimiserError> {
     for stage in &mut pipeline.annotated_stages {
         if let AnnotatedStage::Match { block, block_annotations, .. } = stage {
             optimize_away_statically_unsatisfiable_conjunctions(block.conjunction_mut(), block_annotations);
             prune_redundant_roleplayer_deduplication(block.conjunction_mut(), block_annotations);
-            relation_index_transformation(block.conjunction_mut(), block_annotations, type_manager, snapshot)?;
+            relation_index_transformation(block.conjunction_mut(), block_annotations, thing_manager, snapshot)?;
         }
     }
     Ok(())
