+++ concept/tests/test_statistics.rs
@@ -198,7 +198,7 @@ fn read_statistics(storage: Arc<MVCCStorage<WALClient>>, thing_manager: &ThingMa
 fn create_entity() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
 
     let person_label = Label::build("person", None);
 
@@ -223,7 +223,7 @@ fn create_entity() {
 fn delete_twice() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
 
     let person_label = Label::build("person", None);
 
@@ -253,7 +253,7 @@ fn delete_twice() {
 fn put_has_twice() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
 
     let person_label = Label::build("person", None);
     let name_label = Label::build("name", None);
@@ -308,7 +308,7 @@ fn put_has_twice() {
 fn put_plays() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
 
     let person_label = Label::build("person", None);
     let friendship_label = Label::build("friendship", None);
