+++ concept/tests/test_type.rs
@@ -62,6 +62,7 @@ fn thing_manager(type_manager: Arc<TypeManager>) -> Arc<ThingManager> {
         thing_vertex_generator.clone(),
         type_manager.clone(),
         Arc::new(Statistics::new(DurabilitySequenceNumber::MIN)),
+        false,
     ))
 }
