+++ concept/type_/type_manager/validation/operation_time_validation.rs
@@ -1576,7 +1576,7 @@ impl OperationTimeValidation {
     pub(crate) fn validate_regex_arguments(regex: AnnotationRegex) -> Result<(), Box<SchemaValidationError>> {
         match regex.validate() {
             Ok(_) => Ok(()),
-            Err(err) => Err(Box::new(SchemaValidationError::InvalidRegexArguments { regex, source: err})),
+            Err(err) => Err(Box::new(SchemaValidationError::InvalidRegexArguments { regex, source: err })),
         }
     }
