+++ executor/tests/execute_has.rs
@@ -60,7 +60,7 @@ const NAME_LABEL: Label = Label::new_static("name");
 fn setup_database(storage: &mut Arc<MVCCStorage<WALClient>>) {
     setup_concept_storage(storage);
 
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     let mut snapshot = storage.clone().open_snapshot_write();
 
     let person_type = type_manager.create_entity_type(&mut snapshot, &PERSON_LABEL).unwrap();
@@ -184,7 +184,7 @@ fn traverse_has_unbounded_sorted_from() {
     let entry = builder.finish().unwrap();
 
     let snapshot = storage.clone().open_snapshot_read();
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
 
     let variable_registry = &translation_context.variable_registry;
     let previous_stage_variable_annotations = &BTreeMap::new();
@@ -248,7 +248,7 @@ fn traverse_has_bounded_sorted_from_chain_intersect() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_database(&mut storage);
 
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     // query:
     //   match
     //    $person-1 has name $name;
@@ -390,7 +390,7 @@ fn traverse_has_unbounded_sorted_from_intersect() {
     let entry = builder.finish().unwrap();
 
     let snapshot: ReadSnapshot<WALClient> = storage.clone().open_snapshot_read();
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     let variable_registry = &translation_context.variable_registry;
     let previous_stage_variable_annotations = &BTreeMap::new();
     let block_annotations = infer_types(
@@ -477,7 +477,7 @@ fn traverse_has_unbounded_sorted_to_merged() {
     let entry = builder.finish().unwrap();
 
     let snapshot: ReadSnapshot<WALClient> = storage.clone().open_snapshot_read();
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     let variable_registry = &translation_context.variable_registry;
     let previous_stage_variable_annotations = &BTreeMap::new();
     let block_annotations = infer_types(
@@ -585,7 +585,7 @@ fn traverse_has_reverse_unbounded_sorted_from() {
     let entry = builder.finish().unwrap();
 
     let snapshot: ReadSnapshot<WALClient> = storage.clone().open_snapshot_read();
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     let variable_registry = &translation_context.variable_registry;
     let previous_stage_variable_annotations = &BTreeMap::new();
     let block_annotations = infer_types(
