+++ executor/tests/pipelines.rs
@@ -40,7 +40,7 @@ fn setup_common() -> Context {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
 
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     let function_manager = FunctionManager::new(Arc::new(DefinitionKeyGenerator::new()), None);
     let query_manager = QueryManager::new(None);
     let schema = r#"
@@ -59,7 +59,7 @@ fn setup_common() -> Context {
     snapshot.commit(&mut CommitProfile::DISABLED).unwrap();
 
     // reload to obtain latest vertex generators and statistics entries
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     let query_manager = QueryManager::new(Some(Arc::new(QueryCache::new())));
     Context { _tmp_dir, storage, type_manager, function_manager, query_manager, thing_manager }
 }
