+++ executor/tests/writes.rs
@@ -57,7 +57,7 @@ const NAME_LABEL: Label = Label::new_static("name");
 
 fn setup_schema(storage: Arc<MVCCStorage<WALClient>>) {
     let mut snapshot: WriteSnapshot<WALClient> = storage.clone().open_snapshot_write();
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
 
     let person_type = type_manager.create_entity_type(&mut snapshot, &PERSON_LABEL).unwrap();
     let group_type = type_manager.create_entity_type(&mut snapshot, &GROUP_LABEL).unwrap();
@@ -340,7 +340,7 @@ fn has() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
 
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     setup_schema(storage.clone());
     let snapshot = storage.clone().open_snapshot_write();
     let (_, snapshot) = execute_insert(
@@ -369,7 +369,7 @@ fn test() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
 
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     setup_schema(storage.clone());
 
     let snapshot = storage.clone().open_snapshot_write();
@@ -387,7 +387,7 @@ fn relation() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
 
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     setup_schema(storage.clone());
 
     let snapshot = storage.clone().open_snapshot_write();
@@ -437,7 +437,7 @@ fn relation_with_inferred_roles() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
 
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     setup_schema(storage.clone());
 
     let snapshot = storage.clone().open_snapshot_write();
@@ -487,7 +487,7 @@ fn test_has_with_input_rows() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
 
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     setup_schema(storage.clone());
     let snapshot = storage.clone().open_snapshot_write();
     let (inserted_rows, snapshot) = execute_insert(
@@ -537,7 +537,7 @@ fn delete_has() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
 
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     setup_schema(storage.clone());
     let snapshot = storage.clone().open_snapshot_write();
     let (inserted_rows, snapshot) = execute_insert(
