+++ query/benches/bench_insert_queries_multithreaded.rs
@@ -48,7 +48,7 @@ static NAME_LABEL: OnceLock<Label> = OnceLock::new();
 fn setup_database(storage: &mut Arc<MVCCStorage<WALClient>>) {
     setup_concept_storage(storage);
 
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     let mut snapshot = storage.clone().open_snapshot_write();
     let person_type = type_manager.create_entity_type(&mut snapshot, PERSON_LABEL.get().unwrap()).unwrap();
     let group_type = type_manager.create_entity_type(&mut snapshot, GROUP_LABEL.get().unwrap()).unwrap();
@@ -166,7 +166,7 @@ fn multi_threaded_inserts() {
 
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_database(&mut storage);
-    let (type_manager, thing_manager) = load_managers(storage.clone(), Some(storage.snapshot_watermark()));
+    let (type_manager, thing_manager) = load_managers(storage.clone(), Some(storage.snapshot_watermark()), true);
     let query_manager = QueryManager::new(Some(Arc::new(QueryCache::new())));
     const NUM_THREADS: usize = 32;
     const INTERNAL_ITERS: usize = 1000;
