+++ query/tests/fetch.rs
@@ -61,7 +61,7 @@ fn insert_data(
 fn fetch() {
     let (_tmp_dir, mut storage) = create_core_storage();
     setup_concept_storage(&mut storage);
-    let (type_manager, thing_manager) = load_managers(storage.clone(), None);
+    let (type_manager, thing_manager) = load_managers(storage.clone(), None, false);
     let function_manager = FunctionManager::new(Arc::new(DefinitionKeyGenerator::new()), None);
     define_schema(storage.clone(), type_manager.as_ref(), thing_manager.as_ref(), &function_manager);
     insert_data(
