+++ storage/snapshot/snapshot.rs
@@ -162,7 +162,7 @@ pub trait WritableSnapshot: ReadableSnapshot {
     ) -> Result<ByteArray<BUFFER_VALUE_INLINE>, SnapshotGetError> {
         let keyspace_id = key.keyspace_id();
         let writes = self.operations().writes_in(keyspace_id);
-        match writes.get(key.bytes()) {
+        match writes.writes_get(key.bytes()) {
             Some(Write::Insert { value, .. }) | Some(Write::Put { value, .. }) => Ok(ByteArray::copy(value)),
             Some(Write::Delete) => {
                 Err(SnapshotGetError::ExpectedRequiredKeyToExist { key: StorageKey::Array(key.into_owned_array()) })
@@ -343,8 +343,7 @@ impl<D> ReadableSnapshot for WriteSnapshot<D> {
         key: StorageKeyReference<'_>,
         storage_counters: StorageCounters,
     ) -> Result<Option<ByteArray<INLINE_BYTES>>, SnapshotGetError> {
-        let writes = self.operations().writes_in(key.keyspace_id());
-        match writes.get(key.bytes()) {
+        match self.get_write(key) {
             Some(Write::Insert { value, .. }) | Some(Write::Put { value, .. }) => Ok(Some(ByteArray::copy(value))),
             Some(Write::Delete) => Ok(None),
             None => self
@@ -360,8 +359,7 @@ impl<D> ReadableSnapshot for WriteSnapshot<D> {
         key: StorageKeyReference<'_>,
         storage_counters: StorageCounters,
     ) -> Result<Option<ByteArray<INLINE_BYTES>>, SnapshotGetError> {
-        let writes = self.operations().writes_in(key.keyspace_id());
-        match writes.get(key.bytes()) {
+        match self.get_write(key) {
             Some(Write::Insert { value, .. }) | Some(Write::Put { value, .. }) => Ok(Some(ByteArray::copy(value))),
             Some(Write::Delete) | None => self
                 .storage
@@ -393,7 +391,7 @@ impl<D> ReadableSnapshot for WriteSnapshot<D> {
     }
 
     fn get_write(&self, key: StorageKeyReference<'_>) -> Option<&Write> {
-        self.operations().writes_in(key.keyspace_id()).get_write(key.bytes())
+        self.operations().writes_in(key.keyspace_id()).writes_get(key.bytes())
     }
 
     fn iterate_writes(&self) -> impl Iterator<Item = (StorageKeyArray<BUFFER_KEY_INLINE>, Write)> + '_ {
@@ -501,8 +499,7 @@ impl<D> ReadableSnapshot for SchemaSnapshot<D> {
         key: StorageKeyReference<'_>,
         storage_counters: StorageCounters,
     ) -> Result<Option<ByteArray<INLINE_BYTES>>, SnapshotGetError> {
-        let writes = self.operations().writes_in(key.keyspace_id());
-        match writes.get(key.bytes()) {
+        match self.get_write(key) {
             Some(Write::Insert { value, .. }) | Some(Write::Put { value, .. }) => Ok(Some(ByteArray::copy(value))),
             Some(Write::Delete) => Ok(None),
             None => self
@@ -518,8 +515,7 @@ impl<D> ReadableSnapshot for SchemaSnapshot<D> {
         key: StorageKeyReference<'_>,
         storage_counters: StorageCounters,
     ) -> Result<Option<ByteArray<INLINE_BYTES>>, SnapshotGetError> {
-        let writes = self.operations().writes_in(key.keyspace_id());
-        match writes.get(key.bytes()) {
+        match self.get_write(key) {
             Some(Write::Insert { value, .. }) | Some(Write::Put { value, .. }) => Ok(Some(ByteArray::copy(value))),
             Some(Write::Delete) | None => self
                 .storage
@@ -551,7 +547,7 @@ impl<D> ReadableSnapshot for SchemaSnapshot<D> {
     }
 
     fn get_write(&self, key: StorageKeyReference<'_>) -> Option<&Write> {
-        self.operations().writes_in(key.keyspace_id()).get_write(key.bytes())
+        self.operations().writes_in(key.keyspace_id()).writes_get(key.bytes())
     }
 
     fn iterate_writes(&self) -> impl Iterator<Item = (StorageKeyArray<BUFFER_KEY_INLINE>, Write)> + '_ {
